refactor: extract password management logic into useSecurity hook
- Moved password management state and logic from ChangePasswordCard component to a new useSecurity hook for better separation of concerns and reusability.
- Updated ChangePasswordCard to utilize the new useSecurity hook, simplifying the component structure and enhancing maintainability.
- Ensured that form handling and error management remain intact while improving code organization.
